GPS Drivers submodule.

+++ src/drivers/gps/gps.cpp
@@ -792,7 +792,7 @@ GPS::task_main()
                     /* measure update rate every 5 seconds */
                     if (hrt_absolute_time() - last_rate_measurement > RATE_MEASUREMENT_PERIOD) {
                         float dt = (float)((hrt_absolute_time() - last_rate_measurement)) / 1000000.0f;
-                        _rate = last_rate_count / dt;
+                        _rate = static_cast<float>(last_rate_count) / dt;
                         _rate_rtcm_injection = _last_rate_rtcm_injection_count / dt;
                         last_rate_measurement = hrt_absolute_time();
                         last_rate_count = 0;
